Revert moving dbus_name_owned() to dbusutils.py
Until dbusutils.py ported to GDBus binding. Otherwise it breaks
working on macOS and Windows.
