+++ quodlibet/quodlibet/util/dbusutils.py
@@ -8,8 +8,6 @@
 
 import xml.etree.ElementTree as ET
 
-from gi.repository import GLib, Gio
-
 import dbus
 import dbus.service
 
@@ -337,18 +335,3 @@ class DBusProperty(object):
     @dbus.service.signal(IFACE, signature="sa{sv}as", rel_path_keyword="rel")
     def PropertiesChanged(self, interface, changed, invalidated, rel=""):
         pass
-
-
-def dbus_name_owned(name):
-    """Returns True if the dbus name has an owner"""
-    BUS_DAEMON_NAME = 'org.freedesktop.DBus'
-    BUS_DAEMON_PATH = '/org/freedesktop/DBus'
-    BUS_DAEMON_IFACE = 'org.freedesktop.DBus'
-
-    try:
-        bus = Gio.DBusProxy.new_for_bus_sync(
-            Gio.BusType.SESSION, Gio.DBusProxyFlags.NONE, None,
-            BUS_DAEMON_NAME, BUS_DAEMON_PATH, BUS_DAEMON_IFACE, None)
-        return bus.NameHasOwner('(s)', name)
-    except GLib.Error:
-        return False
